refactor: 상수 폴더 구조 분리
