+++ src/constants/date.js
@@ -1,0 +1,8 @@
+export const DECEMBER = Object.freeze({
+  FIRST: 1,
+  CHRISTMAS: 25,
+  LAST: 31,
+});
+export const WEEKENDS = [1, 2, 8, 9, 15, 16, 22, 23, 29, 30];
+
+export const SPECIAL_DAYS = [3, 10, 17, 24, 25, 31];
