+++ src/constants/event.js
@@ -1,0 +1,15 @@
+export const BADGES = Object.freeze({
+  STAR: "별",
+  TREE: "트리",
+  SANTA: "산타",
+});
+
+export const BENEFIT_PRICE = Object.freeze({
+  HUNDRED: 100,
+  THOUSAND: 1000,
+  FIVE_THOUSAND: 5000,
+  TEN_THOUSAND: 10000,
+  TWENTY_THOUSAND: 20000,
+  YEAR: 2023,
+  GIFT: 120000,
+});
