+++ src/constants/menu.js
@@ -1,0 +1,24 @@
+export const MENUS = Object.freeze({
+  양송이수프: 6000,
+  타파스: 5500,
+  시저샐러드: 8000,
+  티본스테이크: 55000,
+  바비큐립: 54000,
+  해산물파스타: 35000,
+  크리스마스파스타: 25000,
+  초코케이크: 15000,
+  아이스크림: 5000,
+  제로콜라: 3000,
+  레드와인: 60000,
+  샴페인: 25000,
+});
+
+export const MAX_MENU_COUNT = 20;
+
+export const MAIN_MENUS = ["티본스테이크", "바비큐립", "해산물파스타", "크리스마스파스타"];
+
+export const DESSERTS = ["초코케이크", "아이스크림"];
+
+export const DRINKS = ["제로콜라", "레드와인", "샴페인"];
+
+export const ONE_CHAMPAGNE = "샴페인 1개";
